+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # 更新日记【Authz】
 
+## Version 1.0.10 - 2022.7.8
+
+### Added
+
+- 现在`@Roles` `@Perms` `@Certificated`能够对非Mapping进行权限拦截了。
+- 如：对`MyService`中的`test方法`进行登录权限检查
+
+```java
+import cn.omisheep.authz.annotation.Certificated;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+
+@RequestMapping("/method-test")
+public class MethodTestController {
+
+    @Autowired
+    private MyService myService;
+  
+    @GetMapping("/u")
+    public Result u1() {
+      return Result.SUCCESS.data(myService.test());
+    }
+}
+
+@Service
+public class MyService {
+    @Certificated
+    public List<Object> test() {
+      return new ArrayList<Object>();
+    }
+}
+```
+
 ## Version 1.0.9 - 2022.7.7
 
 ### Added
@@ -9,12 +43,12 @@
 ```java
 @GetMapping("/get")
 public Result get(@Decrypt("name") String name){
-    return Result.SUCCESS.data("name",name);
+        return Result.SUCCESS.data("name",name);
 }
 
 @PostMapping("/post")
 public Result post(@Decrypt({"name", "content", "obj.name"}) @RequestBody HashMap<String, Object> map){
-    return Result.SUCCESS.data("map",map);
+        return Result.SUCCESS.data("map",map);
 }
 ```